Fix compilation of x86 image

+++ http.c
@@ -115,7 +115,7 @@ http_get_response(void *buffer, size_t size, size_t rxed, char **msg_in)
 {
 	char *c;
 
-	if (asprintf(&c, "%s%.*s", *msg_in, size * rxed, (char *)buffer) == -1) {
+	if (asprintf(&c, "%s%.*s", *msg_in, (int) (size * rxed), (char *)buffer) == -1) {
 		FREE(*msg_in);
 		return -1;
 	}
